Add support for template strings
