+++ lib/rules/no-auth-route-cst.js
@@ -2,37 +2,6 @@
 /* eslint-disable eslint-plugin/require-meta-type */
 /* eslint-disable eslint-plugin/require-meta-schema */
 module.exports = {
-  // create: (context) => {
-  //   return {
-  //     CallExpression: (node) => {
-  //       const calleeNode = node.callee;
-  //       if (!calleeNode.object) {
-  //         return;
-  //       }
-  //       let rootObject = node.arguments[0];
-  //       while (rootObject.object) {
-  //         rootObject = rootObject.object;
-  //       }
-  //       if (rootObject.name === "routes") {
-  //         const x = context.getSourceCode().getScope(rootObject);
-  //         console.log(x);
-  //       }
-  //       const isError =
-  //         calleeNode.object.name === "router" &&
-  //         (calleeNode.property.name === "push" ||
-  //           calleeNode.property.name === "replace") &&
-  //         (node.arguments[0].value === "/api/auth/login" ||
-  //           node.arguments[0].value === "/api/auth/logout");
-  //       if (isError) {
-  //         context.report({
-  //           node,
-  //           message: "Some error",
-  //         });
-  //       }
-  //     },
-  //   };
-  // },
-
   create: (context) => {
     return {
       CallExpression: (node) => {
@@ -44,9 +13,7 @@ module.exports = {
             (property.name === "push" || property.name === "replace")
           ) {
             const [firstArgument] = node.arguments;
-            const path = firstArgument.value;
-
-            console.log(node.arguments);
+            const path = extractPath(firstArgument);
 
             if (path && path.startsWith("/api/auth/login")) {
               context.report({
@@ -59,4 +26,16 @@ module.exports = {
       },
     };
   },
+};
+
+const extractPath = (argument) => {
+  if (argument.type === "Literal") {
+    return argument.value;
+  }
+
+  if (argument.type === "TemplateLiteral") {
+    return argument.quasis[0].value.raw;
+  }
+
+  throw new Error("Unknown argument type");
 };
