+++ package.json
@@ -19,6 +19,10 @@
     {
       "name": "Henrique Inonhe",
       "email": "henriqueinonhe@gmail.com"
+    },
+    {
+      "name": "Erick Célio",
+      "email": "erickcelio.dev@gmail.com"
     }
   ],
   "main": "./lib/index.js",
@@ -39,7 +43,8 @@
     "eslint-plugin-eslint-plugin": "^5.0.0",
     "eslint-plugin-node": "^11.1.0",
     "mocha": "^10.0.0",
-    "npm-run-all": "^4.1.5"
+    "npm-run-all": "^4.1.5",
+    "prettier": "2.8.8"
   },
   "engines": {
     "node": "^14.17.0 || ^16.0.0 || >= 18.0.0"
